Internal Documentation

Corrections

CatmapInterface.ideal_gas — Function

ideal_gas(energies, catmap_params)

Add thermodynamic correction terms for all gas species using the ideal gas approximation.

CatmapInterface.harmonic_adsorbate — Function

harmonic_adsorbate(energies, catmap_params)

Add thermodynamic correction terms for all adsorbed species using the harmonic adsorbate approximation.

CatmapInterface.hbond_surface_charge_density — Function

hbond_surface_charge_density(
    energies,
    catmap_params,
    σ,
    ϕ_we,
    ϕ,
    local_pH
)

Add electrochemical correction terms for all influenced species using ...

Interface

CatmapInterface.TState — Type

struct TState

A transition state

name::String: Name of the transition state

beta::Float64: Transfer coefficient of the transition state

CatmapInterface.ParsedReaction — Type

struct ParsedReaction

Holds the information of a parsed reaction equation.

educts::Vector{Pair{String, Int64}}: List of educts stored as a pair of the name and stoichiometric factor

products::Vector{Pair{String, Int64}}: List of products stored as a pair of the name and stoichiometric factor

tstate::Union{Nothing, CatmapInterface.TState}: A CatmapInterface.TState if the reaction equation contains a transition state

CatmapInterface.parse_reaction — Function

parse_reaction(r)

Parse a specification of a chemical reaction into a CatmapInterface.ParsedReaction.

Example

julia> CatmapInterface.parse_reaction("CO*_t <-> CO_g + *_t")
CatmapInterface.ParsedReaction(["CO_t" => 1], ["CO_g" => 1, "_t" => 1], nothing)

julia> CatmapInterface.parse_reaction("COOH*_t + H2O_g + ele_g <-> COOH-H2O-ele_t <-> CO*_t + H2O_g + OH_g + *_t; beta=0.5")
CatmapInterface.ParsedReaction(["COOH_t" => 1, "H2O_g" => 1, "ele_g" => 1], ["CO_t" => 1, "H2O_g" => 1, "OH_g" => 1, "_t" => 1], CatmapInterface.TState("COOH-H2O-ele_t", 0.5))

CatmapInterface.parse_reactant_sum — Function

parse_reactant_sum(rs)

Parse a string as a sum of reactants into a list of pairs reactant/stoichiometric factor.

CatmapInterface.parse_energy_table — Function

parse_energy_table(input_file_path)

Parse an energy table into a list of named tuples.

The entries of an energy table must be separated by tabs. The first row of an energy table includes the headers of the columns. The following are recognized (the order of the columns does not matter):

surface_name

site_name

species_name

formation_energy: in eV

bulk_structure

frequencies: list of enclosed with brackets and comma-separated, in cm⁻¹

other_parameters

reference

CatmapInterface.specieslist — Function

specieslist(
    reactions,
    species_defs,
    energy_table,
    surface_name
)

Collect the specifications of all reactants in a list.

CatmapInterface.findspecies — Function

findspecies(species_name, site, species_defs)

Extract specification of a species at a given site from the species_defs of a CatMAP input file.

findspecies(
    species_name,
    energy_table;
    surface_name,
    site_name
)

Extract specification of a species from an energy table.

Reaction Network

CatmapInterface.ratelaw_TS — Function

ratelaw_TS(prefactor, Gf_IS, Gf_TS, T, activprod)

Computes the rate of an elementary reaction that passes through a transition state.

The rate law is based on the Arrhenius relation. The change in Gibbs free energy between the initial and transition state is the needed activation energy. A prefactor and the product of the activities activprod complete the rate law.

CatmapInterface.compute_free_energies! — Function

compute_free_energies!(
    free_energies,
    catmap_params,
    σ,
    ϕ_we,
    ϕ,
    local_pH
)

Compute the Gibbs free energies of all species specified in the catmap_params by applying the specified correction modes.