Interface

CatmapInterface.parse_catmap_input — Function

parse_catmap_input(input_file_path)

Parse the content of a CatMAP input file for creating microkinetic model.

The input file is first interpreted by a Python interpreter and then parsed. The following information is used and must be specified:

rxn_expressions

prefactor_list

species_definitions (must include an entry for every reactant, even if empty)

input_file (including the energy table)

surface_names (currently only the use of exactly one surface is implemented)

bulk_ph

extrapolated_potential

potential_reference_scale

gas_thermo_mode (options: ideal_gas)

adsorbate_thermo_mode (options: harmonic_adsorbate)

electrochemical_thermo_mode (options: simple_electrochemical, hbond_surface_charge_density)

Moreover, the following information is optional and default values exist

adsorbate_interaction_model (options: ideal (default), first_order)

cross_interaction_mode (only used if adsorbate_interaction_model=first_order, options: geometric_mean (default), arithmetic_mean, neglect)

transition_state_cross_interaction_mode (only used if adsorbate_interaction_model=first_order, options: intermediate_state (default), initial_state, final_state, neglect)

interaction_response_function (only used if adsorbate_interaction_mode=first_order, default: smooth_piecewise_linear)

See CatMAP documentation for details.

CatmapInterface.CatmapParams — Type

struct CatmapParams

Main data structure for interfacing CatMAP

reactions::Vector{CatmapInterface.ParsedReaction}: List of parsed reactions

prefactors::Vector{Float64}: Prefactors in the Arrhenius relation specifying the activation energy of the corresponding reactions

species_list::Dict{String, AbstractSpecies}: List of species including all reactants

gas_thermo_mode::Symbol: Mode of the thermodynamical correction to the formation energy of the gases

adsorbate_thermo_mode::Symbol: Mode of the thermodynamical correction to the formation energy of the adsorbates and transition states

electrochemical_thermo_mode::Symbol: Mode of the electrochemical correction to the formation energy of the species

beta_mode::Symbol: Mode for BEP scaling for TStateSpecies

bulk_pH::Float64: pH value in the bulk of the electrolyte

Uref::Float64: Reference potential with respect to the specified potential reference scale

potential_reference_scale::String: Either RHE or SHE

T::Float64: Temperature in the bulk

adsorbate_interaction_params::CatmapInterface.AdsorbateInteractionParams: Parameter specifying the adsorbate interaction model

Species

CatmapInterface.AbstractSpecies — Type

abstract type AbstractSpecies

Abstract supertype of species.

CatmapInterface.GasSpecies — Type

struct GasSpecies <: AbstractSpecies

species_name::String: Name of the species

formation_energy::Float64: Formation energy of the species in joule per mole

pressure::Float64: Partial pressure of the species in pascal

frequencies::Vector{Float64}: Normal modes of vibration of the species in m⁻¹ (wavenumbers)

henry_const::Union{Missing, Float64}: Henry constant used to convert between gas and liquid phase

CatmapInterface.LocalGasSpecies — Type

struct LocalGasSpecies <: AbstractSpecies

species_name::String: Name of the species

formation_energy::Float64: Formation energy of the species in joule per mole

pressure::Float64: Partial pressure or local concentration parameter of the species in pascal

site::String: Site the species adsorbs to / is located at (e.g. "b")

frequencies::Vector{Float64}: Normal modes of vibration of the species in m⁻¹ (wavenumbers)

parent_gas::String: Name of the parent gas species (e.g. "CO_g")

henry_const::Union{Missing, Float64}: Henry constant used to convert between gas and liquid phase

CatmapInterface.AdsorbateSpecies — Type

struct AdsorbateSpecies <: AbstractSpecies

species_name::String: Name of the species

formation_energy::Float64: Formation energy of the species in joule per mole

coverage::Float64: Surface coverage of the species

site::String: Site the species adsorbs to

n_sites::Int64: Number of needed sites for adsorption

surface_name::String: Name of the surface the species adsorbs to

frequencies::Vector{Float64}: Normal modes of vibration of the species in m⁻¹ (wavenumbers)

sigma_params::@NamedTuple{a::Union{Nothing, Float64}, b::Union{Nothing, Float64}}: Parameters specifying the electrochemcial correction to the formation energy due to surface charges: $G_f(U=0, σ) ≈ G_f(U=0,σ=0)+a\cdot σ + b\cdot σ^2$

self_interaction_param::Float64: Parameter used in CatmapInterface.first_order_adsorbate_interaction to specify the formation energy's dependence on the coverage

cross_interaction_params::Dict{String, Float64}: Parameter used in CatmapInterface.first_order_adsorbate_interaction to specify the formation energy's dependence on the coverage of other adsorbates

CatmapInterface.TStateSpecies — Type

struct TStateSpecies <: AbstractSpecies

species_name::String: Name of the species

formation_energy::Union{Nothing, Float64}: Formation energy of the species in joule per mole

n_sites::Int64: Number of needed sites for adsorption

barrier::Union{Nothing, Float64}: Reaction free energy barrier in joule per mole

coverage::Float64: Surface coverage of the species

site::String: Site the species adsorbs to

surface_name::String: Name of the surface the species adsorbs to

frequencies::Vector{Float64}: Normal modes of vibration of the species in m⁻¹

β::Float64: Transfer coefficient of the reaction passing through the specified transition state

between_species::Vector{Pair{String, Int64}}: Specifying the list of (stable) species (with stoichiometric factor) the transition state is between

sigma_params::@NamedTuple{a::Union{Missing, Nothing, Float64}, b::Union{Missing, Nothing, Float64}}: Parameters specifying the electrochemcial correction to the formation energy due to surface charges: $G_f(U=0, σ) ≈ G_f(U=0,σ=0)+a\cdot σ + b\cdot σ^2$

self_interaction_param::Float64: Parameter used in CatmapInterface.first_order_adsorbate_interaction to specify the formation energy's dependence on the coverage

cross_interaction_params::Dict{String, Float64}: Parameter used in CatmapInterface.first_order_adsorbate_interaction to specify the formation energy's dependence on the coverage of other adsorbates

CatmapInterface.FictitiousSpecies — Type

struct FictitiousSpecies <: AbstractSpecies

species_name::String: Name of the species

formation_energy::Float64: Formation energy of the species in joule per mole

pressure::Float64: Partial pressure of the species in pascal

CatmapInterface.SiteSpecies — Type

struct SiteSpecies <: AbstractSpecies

formation_energy::Float64: Formation energy of the site in joule per mole

site_name::String: Name of the site

interaction_response_params::CatmapInterface.InteractionResponseParams: Parameters used to specify the exact form of the interaction_response_function in CatmapInterface.first_order_adsorbate_interaction

Reaction Network

CatmapInterface.create_reaction_network — Function

create_reaction_network(
    catmap_params;
    conserve_pressures,
    symbolic_formation_energies,
    with_free_energies,
    C_gap_val,
    ϕ_pzc_val
)

Create a ReactionSystem for the specified microkinetic model.

For each elementary reaction the CatmapInterface.ratelaw_TS is used. No separate rate equations for the solvent (= H₂O) and the active sites are created but their activities can be specified as parameters. For ficitious gases (OH⁻ and H⁺) and adsorbates the activity coefficients are assumed to be 1. The activity coefficients of the gaseous species can specified as parameters. The thermodynamical corrections to the DFT-data of the formation energies are applied according to the specified modes. New modes can be added by the user by adding a function with the same name to the module.

if conserve_pressures==true, conserve the pressures of the gaseous and fictitious species involved in the heterogeneous reaction network. The pressures of the gaseous and fictitious species are conserved by adding an additional (production/elimination) reaction for each species.

CatmapInterface.liquidize — Function

liquidize(odesys, catmap_params)

Transform the (micro-)kinetic model from surface/gas-reactions to surface/electrolyte-reactions using Henry's law

CatmapInterface.parameter_cache — Function

    parameter_cache(reactionterm, T)

Return a work vector which an hold parameter values of type T. It is indexed by the symbolic parameter names. In the background, it calls the get_tmp function from PreallocationTools.

CatmapInterface.parameter_defaults — Function

 parameter_defaults(odsysys)

Obtain dictionary of default parameter values.

CatmapInterface.generate_function — Function

    generate_function(odesys, species_dict; pdefaults)

Create a ReactionTerm from the right hand side of odesys.

species_dict is a dictionary which for each unknown from odesys contains the species index in a VoronoiFVM.System

pdefaults is a dictionary which holds default parameter values.

CatmapInterface.ReactionTerm — Type

struct ReactionTerm

Callable struct describing a parametrized reaction term built by Catalyst for use in reaction functions in VoronoiFVM.

func::Any: Original ODE function

uidx::Vector{Int64}: Index vector allowing to translate between VoronoiFVM species numbers and species indices in ODE function func. For a given u from VoronoiFVM, view(u, pidx) can be passed to func.

pdict::Dict{Symbol, Int64}: Dictionary allowing to index a parameter in a parameter buffer.

pdefaults::Union{Nothing, Dict{Symbol, Float64}}: Optional default values for parameters. Can be used to set fixed parameter values which are not dependent on unknowns.

cache::Any: PreallocationTools.DiffCache providing workspace for parameters.

CatmapInterface.ReactionTermParameterCache — Type

struct ReactionTermParameterCache

Struct which holds a parameter workspace, indexed by symbolic parameter names.

API for direct handling of ODE functions

CatmapInterface.unknown_indexes — Function

unknown_indexes(odesys, species_dict::Dict{String, Int}; symb2name)

For the symbolic names of unknowns in odesys, return the species indices given in species_dict.

symb2name is a function which takes a symbolic object and converts it into a string serving as the key in species_dict. Default is string.

CatmapInterface.parameter_indexes — Function

parameter_indexes(odesys, params_dict::Dict{String, Int}; symb2name)

For the symbolic names of parameters in odesys, return the indices given in params_dict.

symb2name is a function which takes a symbolic object and converts it into a string serving as the key in params_dict. Default is string.

CatmapInterface.parameter_dict — Function

parameter_dict(odesys)

Return a Dict{Symbol, Int} which allows to retrieve the index of a parameter in a vector of values.

CatmapInterface.init_params! — Function

init_params!(ps, odesys)

Populate parameter vector ps with default parameter values from odesys.

CatmapInterface.default_params — Function

default_params(odesys)

Return parameter vector for odesys initialized with default parameter values from ModelingToolkit defaults.

Package

CatmapInterface — Module

CatmapInterface

CatmapInterface: CatMap based microkinetic models in Julia

CatmapInterface implements the functionality of a subset of the Python package CatMAP that is used in the Python package CatINT. Whereas CatINT uses an iterative approach for combining the Poisson-Nernst-Planck transport model with a microkinetic model of the surface reactions at the electrode, the output of CatmapInterface can be directly plugged into the functionality of LiquidElectrolytes to solve the coupled system.

Installation

Version >=0.4

Starting with version 0.4, the package can be installed using the Julia package manager from the Julia General Registry

Older versions

Package versions up to v0.3.1 are registered in the julia package registry https://github.com/j-fu/PackageNursery To add the registry (needed only once), and to install the package, from the Julia REPL, type ] to enter the Pkg REPL mode and run:

pkg> registry add https://github.com/j-fu/PackageNursery
pkg> add https://github.com/ElcatFVM/CatmapInterface.jl

Please be aware that adding a registry to your Julia installation requires to trust the registry maintainer for handling things in a correct way. In particular, the registry should not register higher versions of packages which are already registered in the Julia General Registry. One can check this by visiting the above mentionend github repository URL and inspecting the contents.