Public Documentation

General Remarks

Interface

CatmapInterface.parse_catmap_input — Function

parse_catmap_input(input_file_path)

Parse the content of a CatMAP input file needed for a Poisson-Nernst-Planck model.

The input file is first interpreted by a Python interpreter and then parsed. The following information is used and must be specified:

rxn_expressions

prefactor_list

species_definitions (including ...)

input_file (including the energy table)

surface_names

bulk_ph

Upzc

potentialreferencescale

gasthermomode

adsorbatethermomode

electrochemicalthermomode

See CatMAP documentation for details.

CatmapInterface.CatmapParams — Type

struct CatmapParams

Main data structure for interfacing CatMAP

reactions::Vector{CatmapInterface.ParsedReaction}: List of parsed reactions

prefactors::Vector{Float64}: Prefactors in the Arrhenius relation specifying the activation energy of the corresponding reactions

species_list::Dict{String, AbstractSpecies}: List of species including all reactants

gas_thermo_mode::Symbol: Mode of the thermodynamical correction to the formation energy of the gases

adsorbate_thermo_mode::Symbol: Mode of the thermodynamical correction to the formation energy of the adsorbates and transition states

electrochemical_thermo_mode::Symbol: Mode of the electrochemical correction to the formation energy of the species

bulk_pH::Float64: pH value in the bulk of the electrolyte

Upzc::Float64: Potential of zero charge with respect to the specified potential reference scale

potential_reference_scale::String: Either RHE or SHE

T::Float64: Temperature in the bulk

Species

CatmapInterface.AbstractSpecies — Type

abstract type AbstractSpecies

Abstract supertype of species.

CatmapInterface.GasSpecies — Type

struct GasSpecies <: AbstractSpecies

species_name::String: Name of the species

formation_energy::Float64: Formation energy of the species in joule per mole

pressure::Float64: Partial pressure of the species in pascal

frequencies::Vector{Float64}: Normal modes of vibration of the species in m⁻¹

henry_const::Union{Missing, Float64}:

CatmapInterface.AdsorbateSpecies — Type

struct AdsorbateSpecies <: AbstractSpecies

species_name::String: Name of the species

formation_energy::Float64: Formation energy of the species in joule per mole

coverage::Float64: Surface coverage of the species

site::String: Site the species adsorbs to

surface_name::String: Name of the surface the species adsorbs to

frequencies::Vector{Float64}: Normal modes of vibration of the species in m⁻¹

sigma_params::NamedTuple{(:a, :b), Tuple{Float64, Float64}}: Parameters specifying the electrochemcial correction to the formation energy due to surface charges: $G_f(U=0, σ) ≈ G_f(U=0,σ=0)+a\cdot σ + b\cdot σ^2$

CatmapInterface.TStateSpecies — Type

struct TStateSpecies <: AbstractSpecies

species_name::String: Name of the species

formation_energy::Float64: Formation energy of the species in joule per mole

coverage::Float64: Surface coverage of the species

site::String: Site the species adsorbs to

surface_name::String: Name of the surface the species adsorbs to

frequencies::Vector{Float64}: Normal modes of vibration of the species in m⁻¹

sigma_params::NamedTuple{(:a, :b), Tuple{Float64, Float64}}: Parameters specifying the electrochemcial correction to the formation energy due to surface charges: $G_f(U=0, σ) ≈ G_f(U=0,σ=0)+a\cdot σ + b\cdot σ^2$

β::Float64: Transfer coefficient of the reaction passing through the specified transition state

between_species::Vector{String}: Specifying the list of (stable) species the transition state is between

CatmapInterface.FictiousSpecies — Type

struct FictiousSpecies <: AbstractSpecies

species_name::String: Name of the species

formation_energy::Float64: Formation energy of the species in joule per mole

pressure::Float64: Partial pressure of the species in pascal

CatmapInterface.SiteSpecies — Type

struct SiteSpecies <: AbstractSpecies

formation_energy::Float64: Formation energy of the site in joule per mole

site_name::String: Name of the site

Reaction Network

CatmapInterface.create_reaction_network — Function

create_reaction_network(catmap_params)

Create a ReactionSystem for the specified microkinetic model.

For each elementary reaction the CatmapInterface.ratelaw_TS is used. No separate rate equations for the solvent (= H₂O) and the active sites are created but their activities can be specified as parameters. For ficitious gases (OH⁻ and H⁺) and adsorbates the activity coefficients are assumed to be 1. The activity coefficients of the gaseous species can specified as parameters. The thermodynamical corrections to the DFT-data of the formation energies are applied according to the specified modes. New modes can be added by the user by adding a function with the same name to the module.

CatmapInterface.generate_function — Function

generate_function(rn; dvs, ps)

Generate a mutating function from a ReactionSystem that computes the concentration fluxes due to the reaction.

generate_function(sys; dvs, ps)

Generate a mutating function from a ODESystem that computes the concentration fluxes due to the reaction.

CatmapInterface.liquidize — Function

liquidize(odesys, catmap_params)

Transform the (micro-)kinetic model from surface/gas-reactions to surface/electrolyte-reactions using Henry's law.