Public Documentation

General Remarks

Interface

CatmapInterface.parse_catmap_input — Function

parse_catmap_input(input_file_path)

Parse the content of a CatMAP input file for creating microkinetic model.

The input file is first interpreted by a Python interpreter and then parsed. The following information is used and must be specified:

rxn_expressions

prefactor_list

species_definitions (must include an entry for every reactant, even if empty)

input_file (including the energy table)

surface_names (currently only the use of exactly one surface is implemented)

bulk_ph

extrapolated_potential

potential_reference_scale

gas_thermo_mode (options: ideal_gas)

adsorbate_thermo_mode (options: harmonic_adsorbate)

electrochemical_thermo_mode (options: simple_electrochemical, hbond_surface_charge_density)

Moreover, the following information is optional and default values exist

adsorbate_interaction_model (options: ideal (default), first_order)

cross_interaction_mode (only used if adsorbate_interaction_model=first_order, options: geometric_mean (default), arithmetic_mean, neglect)

transition_state_cross_interaction_mode (only used if adsorbate_interaction_model=first_order, options: intermediate_state (default), initial_state, final_state, neglect)

interaction_response_function (only used if adsorbate_interaction_mode=first_order, default: smooth_piecewise_linear)

See CatMAP documentation for details.

CatmapInterface.CatmapParams — Type

struct CatmapParams

Main data structure for interfacing CatMAP

reactions::Vector{CatmapInterface.ParsedReaction}: List of parsed reactions

prefactors::Vector{Float64}: Prefactors in the Arrhenius relation specifying the activation energy of the corresponding reactions

species_list::Dict{String, AbstractSpecies}: List of species including all reactants

gas_thermo_mode::Symbol: Mode of the thermodynamical correction to the formation energy of the gases

adsorbate_thermo_mode::Symbol: Mode of the thermodynamical correction to the formation energy of the adsorbates and transition states

electrochemical_thermo_mode::Symbol: Mode of the electrochemical correction to the formation energy of the species

bulk_pH::Float64: pH value in the bulk of the electrolyte

Uref::Float64: Reference potential with respect to the specified potential reference scale

potential_reference_scale::String: Either RHE or SHE

T::Float64: Temperature in the bulk

adsorbate_interaction_params::CatmapInterface.AdsorbateInteractionParams: Parameter specifying the adsorbate interaction model

Species

CatmapInterface.AbstractSpecies — Type

abstract type AbstractSpecies

Abstract supertype of species.

CatmapInterface.GasSpecies — Type

struct GasSpecies <: AbstractSpecies

species_name::String: Name of the species

formation_energy::Float64: Formation energy of the species in joule per mole

pressure::Float64: Partial pressure of the species in pascal

frequencies::Vector{Float64}: Normal modes of vibration of the species in m⁻¹

henry_const::Union{Missing, Float64}: Henry constant used to convert between gas and liquid phase

CatmapInterface.AdsorbateSpecies — Type

struct AdsorbateSpecies <: AbstractSpecies

species_name::String: Name of the species

formation_energy::Float64: Formation energy of the species in joule per mole

coverage::Float64: Surface coverage of the species

site::String: Site the species adsorbs to

surface_name::String: Name of the surface the species adsorbs to

frequencies::Vector{Float64}: Normal modes of vibration of the species in m⁻¹

sigma_params::NamedTuple{(:a, :b), Tuple{Union{Nothing, Float64}, Union{Nothing, Float64}}}: Parameters specifying the electrochemcial correction to the formation energy due to surface charges: $G_f(U=0, σ) ≈ G_f(U=0,σ=0)+a\cdot σ + b\cdot σ^2$

self_interaction_param::Float64: Parameter used in CatmapInterface.first_order_adsorbate_interaction to specify the formation energy's dependence on the coverage

cross_interaction_params::Dict{String, Float64}: Parameter used in CatmapInterface.first_order_adsorbate_interaction to specify the formation energy's dependence on the coverage of other adsorbates

CatmapInterface.TStateSpecies — Type

struct TStateSpecies <: AbstractSpecies

species_name::String: Name of the species

formation_energy::Float64: Formation energy of the species in joule per mole

coverage::Float64: Surface coverage of the species

site::String: Site the species adsorbs to

surface_name::String: Name of the surface the species adsorbs to

frequencies::Vector{Float64}: Normal modes of vibration of the species in m⁻¹

β::Float64: Transfer coefficient of the reaction passing through the specified transition state

between_species::Vector{Pair{String, Int64}}: Specifying the list of (stable) species (with stoichiometric factor) the transition state is between

sigma_params::NamedTuple{(:a, :b), Tuple{Union{Nothing, Float64}, Union{Nothing, Float64}}}: Parameters specifying the electrochemcial correction to the formation energy due to surface charges: $G_f(U=0, σ) ≈ G_f(U=0,σ=0)+a\cdot σ + b\cdot σ^2$

cross_interaction_params::Dict{String, Float64}: Parameter used in CatmapInterface.first_order_adsorbate_interaction to specify the formation energy's dependence on the coverage of other adsorbates

CatmapInterface.FictiousSpecies — Type

struct FictiousSpecies <: AbstractSpecies

species_name::String: Name of the species

formation_energy::Float64: Formation energy of the species in joule per mole

pressure::Float64: Partial pressure of the species in pascal

CatmapInterface.SiteSpecies — Type

struct SiteSpecies <: AbstractSpecies

formation_energy::Float64: Formation energy of the site in joule per mole

site_name::String: Name of the site

interaction_response_params::CatmapInterface.InteractionResponseParams: Parameters used to specify the exact form of the interaction_response_function in CatmapInterface.first_order_adsorbate_interaction

Reaction Network

CatmapInterface.create_reaction_network — Function

create_reaction_network(catmap_params)

Create a ReactionSystem for the specified microkinetic model.

For each elementary reaction the CatmapInterface.ratelaw_TS is used. No separate rate equations for the solvent (= H₂O) and the active sites are created but their activities can be specified as parameters. For ficitious gases (OH⁻ and H⁺) and adsorbates the activity coefficients are assumed to be 1. The activity coefficients of the gaseous species can specified as parameters. The thermodynamical corrections to the DFT-data of the formation energies are applied according to the specified modes. New modes can be added by the user by adding a function with the same name to the module.

CatmapInterface.generate_function — Function

generate_function(rn; dvs, ps)

Generate a mutating function from a ReactionSystem that computes the concentration fluxes due to the reaction.

generate_function(sys; dvs, ps)

Generate a mutating function from a ODESystem that computes the concentration fluxes due to the reaction.

CatmapInterface.liquidize — Function

liquidize(odesys, catmap_params)

Transform the (micro-)kinetic model from surface/gas-reactions to surface/electrolyte-reactions using Henry's law.